refactor(materializer): extract entry-materializers (file #3)
Lift the entry-level materializers (operation/transcript/question-interaction ->
scene entry: operationSnapshotToToolCall, collectChildOperations,
materializeOperationEntry, materializeMissingToolEntry, materializeTranscriptEntry,
questionInteractionToSceneEntry + interactionSceneEntryId) into entry-materializers.ts,
and prune the now-unused imports. Net 74/74, check 0.
